La funcionalidad para redimencionar imagenes y validar el pero ya esta lista

+++ controllers/administrativo/DocenteController.php
@@ -164,23 +164,22 @@ class DocenteController
     {
         $file = $_FILES['foto_perfil'];
         $filename = $file['name'];
-        $mimetype = $file['type'];
-        if ($mimetype == "image/jpg" || $mimetype == "image/jpeg" || $mimetype == "image/png" || $mimetype == "image/gif") {
-
-            if (!is_dir('photos/docentes/')) {
-                mkdir('photos/docentes/', 0777, true);
+        $tmp = $file['tmp_name'];
+        # Redimencionar y validara que la imagen no pese mas de 2 MB
+        $intenvencion = Utils::intevenirImagen('photos/docentes/', $filename, $tmp);
+        if ($intenvencion) {
+            $id = $_POST['docente'];
+            $foto_actual = $_POST['foto_actual'];
+            $new_photo = new Docente();
+            $new_photo->setId($id);
+            $new_photo->setImg($filename);
+            $resultadoF = $new_photo->uptdateImgPerfil();
+            if ($resultadoF) {
+                unlink('photos/docentes/' . $foto_actual);
             }
-            move_uploaded_file($file['tmp_name'], 'photos/docentes/' . $filename);
-        }
-        $id = $_POST['docente'];
-        $foto = $filename;
-        $foto_actual = $_POST['foto_actual'];
-        $new_photo = new Docente();
-        $new_photo->setId($id);
-        $new_photo->setImg($foto);
-        $resultadoF = $new_photo->uptdateImgPerfil();
-        if ($resultadoF) {
-            unlink('photos/docentes/' . $foto_actual);
+            Utils::alertas($resultadoF, 'Foto de perfil actualizada con éxito', 'algo salió mal, intentelo de nuevo');
+        } else {
+            Utils::alertas($intenvencion, '', 'La imagen es muy pesada, tiene peque pesar menos de 2 MB');
         }
         header("Location: " . base_url . 'Docente/perfilDocente&id=' . $_POST['docente']);
     }
